AP-36 Temporarily Remove telescope from being a dev dependency
The sandbox release build is currently failing because it cant find telescope, suggesting the initial setup made it required and it's typical installed as a dev dependency which wont be part of the release.

This is a temporal fix to hopefully save the stress of finding where its been set.