AP-88 Refactor Baloon createUserForAuth method to use entire payload
Since the created user for baloon needs to have extra information that's present within the rest of the payload, it makes sense to have it access the entire payload instead of being restricted to the JWT token only.

This also eliminates the need for duplicate calls to decoding the token outside the method where needed.

- move the jwt decoding to happen within the method itself
- remove duplicate jwt decoding calls from the actions calling the method