AP-88 Create action to apply access right limits on a list query
An action which would fetch the access rights of specified user, and apply the rules as baloon expects on a query which is meant to return a list. Currently works only for customers listing query.

The company limitation access right rule uses the Ids of the insurers as they exist on baloon's DB to find them and limit the result to customers who have those insurers.