AP-88 Implement access right restriction on Policies endpoint
Turns out the frontend was using the policies list endpoint on the client list page instead of the customers list endpoint.

This is to implement the same logic that was done to customer list on the policy list endpoint as well.

This also required modifying the ApplyAccessRightsToListQuery action to use the proper relationship name.